parallelepipeds/requirements.txt
numpy==1.23.1
torch==1.13.1
scipy==1.9.0
matplotlib==3.5.2
tqdm==4.64.0
seaborn==0.12.2
absl-py